refactor: Moved indecision app (jsx) to playground dir.
The current indecision app is all in jsx, there are no components
that make it up. As a result, it was moved to the playground
directory, as the app.js is now going to start forming the app
using components.
