+++ src/index.js
@@ -10,6 +10,8 @@ var template = (
   </React.StrictMode>
 )
 
+var appRoot = document.getElementById('root')
+ReactDOM.render(template, appRoot);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
